refactor: reorganize foreman file structure into clients/, controllers/, models/
- Move db-client.ts and github.ts to src/foreman/clients/ (external service adapters)
- Move admin-ws.ts to src/foreman/controllers/ (WebSocket server belongs with wss.ts)
- Move event-queue.ts to src/foreman/models/ (data structure used by task-manager)
- Eliminate event-fmt.ts by inlining fmtEventDetails as a private static method on
  WebhookEvent.format(), and updating wss.ts to call evt.format() directly

After these moves, src/foreman/ contains only index.ts and the three subdirectories.

Closes #675
