+++ CLAUDE.md
@@ -10,9 +10,10 @@ Two independent processes: **foreman** (server) and **worker** (agent). They com
 
 MVC structure. Models own state; controllers handle external inputs.
 
-- **Models** (`models/`) — `ActiveRecord` (`active-record.ts`) is the abstract base class for all DB-backed models; it provides `select()`, `get()`, `getBy()`, `list()`, `insert()`, `update()`, and `delete()` using `db.from(tableName)` with `as any` casts since the table name is a runtime string. `Task` is the active-record for tasks; `task.toAssignmentPayload()` returns the issue object for the `task_assigned` wire message. `WebhookEvent` is the active-record for incoming GitHub webhook events (`webhook_events` table); `WebhookEvent.fromIncoming()` builds an in-memory instance and `toWorkerPayload()` maps `eventName` → `name` for the wire protocol. `ForemanMessage` is the active-record for foreman↔worker messages (`foreman_messages` table); `ForemanMessage.buildSummary()` is the single source of truth for log entry summaries. `queryActivityLog()` (in `activity-log.ts`) merges both tables by timestamp. `TaskManager` owns ephemeral in-memory state only (event queues, branch mappings, blocker state) and encapsulates all issue/PR event lifecycle logic — `handleIssueLabeledEvent()`, `handleIssueBodyEditedEvent()`, `handlePrOpenedEvent()`, `handlePrClosedEvent()`, `getTaskForCheckEvent()`. It also exposes `assignIdleWorkers()` (assigns all pending tasks to idle workers, returns `Promise<AssignOutcome[]>`) and `startDepsLoad()` (fire-and-forget dep fetch; emits `"deps_loaded"` on completion so the controller can subscribe once and call `assignWork()`). `Worker` is the in-memory model for connected workers (module-level registry, static finders, `Worker._reset()` for test isolation) — it does not extend `ActiveRecord` since workers are not DB-backed.
+- **Models** (`models/`) — `ActiveRecord` (`active-record.ts`) is the abstract base class for all DB-backed models; it provides `select()`, `get()`, `getBy()`, `list()`, `insert()`, `update()`, and `delete()` using `db.from(tableName)` with `as any` casts since the table name is a runtime string. `Task` is the active-record for tasks; `task.toAssignmentPayload()` returns the issue object for the `task_assigned` wire message. `WebhookEvent` is the active-record for incoming GitHub webhook events (`webhook_events` table); `WebhookEvent.fromIncoming()` builds an in-memory instance and `toWorkerPayload()` maps `eventName` → `name` for the wire protocol; `WebhookEvent.format()` returns a human-readable admin-dashboard string (event formatting helpers are private static methods on this class, kept here so the foreman has zero imports from agent/worker TUI code). `ForemanMessage` is the active-record for foreman↔worker messages (`foreman_messages` table); `ForemanMessage.buildSummary()` is the single source of truth for log entry summaries. `queryActivityLog()` (in `activity-log.ts`) merges both tables by timestamp. `EventQueue` (`event-queue.ts`) buffers GitHub webhook events per task for workers that aren't yet connected. `TaskManager` owns ephemeral in-memory state only (event queues, branch mappings, blocker state) and encapsulates all issue/PR event lifecycle logic — `handleIssueLabeledEvent()`, `handleIssueBodyEditedEvent()`, `handlePrOpenedEvent()`, `handlePrClosedEvent()`, `getTaskForCheckEvent()`. It also exposes `assignIdleWorkers()` (assigns all pending tasks to idle workers, returns `Promise<AssignOutcome[]>`) and `startDepsLoad()` (fire-and-forget dep fetch; emits `"deps_loaded"` on completion so the controller can subscribe once and call `assignWork()`). `Worker` is the in-memory model for connected workers (module-level registry, static finders, `Worker._reset()` for test isolation) — it does not extend `ActiveRecord` since workers are not DB-backed.
 - **Controllers** (`controllers/`) — `http-server.ts` handles webhooks + REST + SPA. `wss.ts` is the protocol layer: handles both directions of the worker WebSocket channel and webhook routing. `ForemanWss` is a thin controller — it parses webhook payloads, calls one `TaskManager` method per event action, and sends wire messages. Instance methods: routing (`routePrEvent`, `routeCheckEvent`, `routeIssueEvent`, `forwardEvent`, `assignWork`), hello handling (`handleBusyHello`, `handleIdleHello`), and helpers (`sendMsg`, `workerLog`). No standalone functions or deps objects exist; everything is on the class.
-- **Infrastructure** — `db-client.ts` wires the shared Supabase client and exports `DbRow<T>`, a helper type that makes `id` optional for in-memory (unsaved) model instances. `admin-ws.ts` broadcasts to the admin dashboard. `github.ts` wraps GitHub API calls.
+- **Clients** (`clients/`) — `db-client.ts` wires the shared Supabase client and exports `DbRow<T>`, a helper type that makes `id` optional for in-memory (unsaved) model instances. `github.ts` wraps GitHub API calls (Supabase and GitHub API adapters live here).
+- `admin-ws.ts` (in `controllers/`) broadcasts to the admin dashboard WebSocket — a WebSocket server that belongs alongside `wss.ts` in the controllers layer.
 
 ### Agent/Worker (`src/agent/`)
