+++ src/foreman/clients/db-client.ts
@@ -1,5 +1,5 @@
 import type { SupabaseClient } from "@supabase/supabase-js";
-import type { Database } from "../database.types.js";
+import type { Database } from "../../database.types.js";
 
 export let db: SupabaseClient<Database>;
